krihelinator v0.0.1 Krihelinator.Background.Poller

A task that loop through all github repositories, using the github API, and send jobs to the StatsScraper.

When there are no more repos to fetch, it simply dies, and the supervisor restarts it.

Functions

next_path(repos)

process_repo(repo)

process_repos(url)

Fetch repos with the github API, and send them to the StatsGetter.

start_link()