krihelinator v0.0.1 Krihelinator.Background.StatsScraper

Poolboy worker that scrape statistics for each new repo. Than, push the info down to the DataHandler sink GenServer.

Functions

general_extractor(floki, css_selector, regex_pattern)

Extracts information from the “floki-parsed” html using css selectors and regex matching on the resulting text.

handle_failure(httpoison_response, repo_name)

When HTTPoison fails, log the failure in the most indicative way.

parse(body)

Use floki to parse the page and return a map for that repo.

process(repo_name)

Fetch statistics about a repository, using pool workers, and send it down to the persistance layer.

start_link(list)

to_one_line(text)

Remove new lines and extra spaces from strings.

Example:

iex> import Krihelinator.Background.StatsScraper, only: [to_one_line: 1] iex> to_one_line(“hello\nworld”) “hello world” iex> to_one_line(“ too\n many spaces “) “too many spaces”