krihelinator v0.0.1 Krihelinator.Background.TrendingPoller

Every :trending_poller_period, scrape the github trending page for interesting, active, projects.

Functions

parse_description(floki_item)

Parse the repo description from the floki item, or :nil if doesn’t exist.

parse_item(floki_item)

Parse single floki item repo to “name” and “description”.

parse_name(floki_item)

Parse the repo name (user/repo) from the repo floki item.

poll()

Poll the github trending page, send each project to the scrapers to process.

start_link()