krihelinator v0.0.1 Krihelinator.GithubRepo

Functions

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

set_krihelimeter(changeset)

Use the existing data, and the expected changes, to calculate and set the krihelimeter.