krihelinator v0.0.1 Krihelinator.Periodic

Every :periodic_schedule do the following:

Scrape the github trending page for interesting, active, projects.

Run the DB cleaner.

Pass all of the repos from the DB through the parser again, to update stats.

Functions

clean_db()

Keep only the top :max_repos_to_keep in the DB. Delete the rest.

reanalyse_existing_repos()

Re-analyse each project in the DB.

scrape_and_persist_repos(repos)

Helper to stream repos through Pipeline.StatsScraper -> Pipeline.DataHandler.

scrape_trending()

Scrape the github trending page, scrape the pulse page for each project and persist.

start_link()