krihelinator v0.0.1 Krihelinator.Periodic

Every :periodic_schedule do the following:

Scrape the github trending page for interesting, active, projects.

Run the DB cleaner.

Pass all of the repos from the DB through the parser again, to update stats.

Functions

clean_db()

Keep only the top :max_repos_to_keep in the DB. Delete the rest.

existing_repos_to_scrape()

handle_scraped(repo)

Decide what to do with the scraped data. Specific errors might trigger save, other deletes, or ignores.

reschedule_work()

Schedule the next run in :periodic_schedule milliseconds.

scrape_trending()

Scrape the github trending page and return stream of repos to scrape.

start_link()